Refactor Dashboard and Login pages to use dedicated view components; implement data fetching and form handling in DashboardView and LoginView respectively.
